APP_example/lora_sd/requirements.txt
torch
torchvision
peft
datasets
git+https://github.com/huggingface/accelerate
git+https://github.com/huggingface/diffusers
git+https://github.com/huggingface/transformers
tqdm
ftfy
Pillow
bitsandbytes
gradio